requirements.txt
streamlit>=1.22.0
boto3>=1.26.0
Pillow>=9.5.0
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.28.0 